+++ wxc_sdk/cdr/__init__.py
@@ -1,6 +1,7 @@
 """
 CDR API
 """
+import re
 from collections.abc import Generator
 from dataclasses import dataclass
 from datetime import datetime, timedelta
@@ -85,6 +86,7 @@ class CDRRelatedReason(str, Enum):
     voice_xml_script_termination = 'VoiceXMLScriptTermination'
     call_forward_no_answer = 'CallForwardNoAnswer'
     anywhere_location = 'AnywhereLocation'
+    call_retrieve = 'CallRetrieve'
 
 
 class CDRUserType(str, Enum):
@@ -101,6 +103,26 @@ class CDRUserType(str, Enum):
     voice_xml = 'VoiceXML'
     route_point = 'RoutePoint'
     virtual_line = 'VirtualLine'
+    place = 'Place'
+
+
+CAMEL_RE = re.compile(r'\b ?(\w+)')
+
+
+def camel(name: str) -> str:
+    """
+    get a camel case name for a field name in a CDR
+    Example: Answer time -> answerTime
+
+    :meta private:
+    """
+    def replacement(m)->str:
+        r = m.group(1).lower().capitalize()
+        return r
+
+    r, _ = re.subn(CAMEL_RE, replacement, name)
+    r = f'{r[0].lower()}{r[1:]}'
+    return r
 
 
 class CDR(ApiModel):
@@ -108,86 +130,84 @@ class CDR(ApiModel):
     @root_validator(pre=True)
     def force_none(cls, values: dict):
         """
-        Pop all empty strings so that they get caught by Optional[]
+        Pop all empty strings so that they get caught by Optional[] and convert keys to camelCase
         :param values:
         :return:
         """
-        values = {k: v for k,v in values.items() if v != ''}
-        # empty_attrs = [k for k, v in values.items() if v == '' and k.endswith('time')]
-        # for k in empty_attrs:
-        #     values.pop(k)
+        # get rid of all empty values and convert to camelCase
+        values = {camel(k): v for k, v in values.items() if v != ''}
         return values
 
+    #: This is the start time of the call, the answer time may be slightly after this. Time is in UTC.
+    start_time: Optional[datetime]
     #: The time the call was answered. Time is in UTC.
-    answer_time: Optional[datetime] = Field(alias='Answer time')
+    answer_time: Optional[datetime]
+    #: The length of the call in seconds.
+    duration: Optional[int]
     #: Whether the call leg was answered. For example, in a hunt group case, some legs will be unanswered,
     # and one will be answered.
-    answered: Optional[bool] = Field(alias='Answered')
+    answered: Optional[bool]
     #: Whether the call was inbound or outbound. The possible values are:
-    direction: Optional[Union[CDRDirection, str]] = Field(alias='Direction')
+    direction: Optional[Union[CDRDirection, str]]
     #: For incoming calls, the calling line ID of the user. For outgoing calls, it's the calling line ID of the
     #: called party.
-    called_line_id: Optional[str] = Field(alias='Called line ID')
+    called_line_id: Optional[str]
     #: SIP Call ID used to identify the call. You can share the Call ID with Cisco TAC to help them pinpoint a call
     # if necessary.
-    call_id: Optional[str] = Field(alias='Call ID')
+    call_id: Optional[str]
     #: For incoming calls, the calling line ID of the calling party. For outgoing calls, it's the calling line ID of
     # the user.
-    calling_line_id: Optional[str] = Field(alias='Calling line ID')
-    #: This is the start time of the call, the answer time may be slightly after this. Time is in UTC.
-    start_time: Optional[datetime] = Field(alias='Start time')
+    calling_line_id: Optional[str]
     #: Type of call. For example:
-    call_type: Optional[CDRCallType] = Field(alias='Call type')
+    call_type: Optional[CDRCallType]
     #: The type of client that the user (creating this record) is using to make or receive the call. For example:
-    client_type: Optional[CDRClientType] = Field(alias='Client type')
+    client_type: Optional[CDRClientType]
     #: The version of the client that the user (creating this record) is using to make or receive the call.
-    client_version: Optional[str] = Field(alias='Client version')
+    client_version: Optional[str]
     #: Correlation ID to tie together multiple call legs of the same call session.
-    correlation_id: Optional[str] = Field(alias='Correlation ID')
+    correlation_id: Optional[str]
     #: The country code of the dialed number. This is only populated for international calls.
-    international_country: Optional[str] = Field(alias='International country')
+    international_country: Optional[str]
     #: The Session ID comprises a Universally Unique Identifier (UUID) for each user-agent participating in a call. It
     #: can be used for end-to-end tracking of a SIP session in IP-based multimedia communication. Each call consists of
     #: two UUIDs known as Local Session ID and Remote Session ID.
     #:   * The Local SessionID is generated from the Originating user agent.
-    local_session_id: Optional[str] = Field(alias='Local SessionID')
+    local_session_id: Optional[str] = Field(alias='localSessionid')
     #: The MAC address of the device, if known.
-    device_mac: Optional[str] = Field(alias='Device MAC')
-    #: The length of the call in seconds.
-    duration: Optional[int] = Field(alias='Duration')
+    device_mac: Optional[str]
     #: Inbound trunk may be presented in Originating and Terminating records.
-    inbound_trunk: Optional[str] = Field(alias='Inbound trunk')
+    inbound_trunk: Optional[str]
     #: A unique identifier for the organization that made the call. This is a unique identifier across Cisco.
-    org_uuid: Optional[str] = Field(alias='Org UUID')
+    org_uuid: Optional[str]
     #: Populated for calls that transfer, hold, wait, and so on. For example:
-    original_reason: Optional[CDROriginalReason] = Field(alias='Original reason')
+    original_reason: Optional[CDROriginalReason]
     #: The operating system that the app was running on, if available.
-    os_type: Optional[str] = Field(alias='OS type')
+    os_type: Optional[str]
     #: Outbound trunk may be presented in Originating and Terminating records.
-    outbound_trunk: Optional[str] = Field(alias='Outbound trunk')
+    outbound_trunk: Optional[str]
     #: Populated for calls that transfer, hold, wait, and so on. For example:
-    redirect_reason: Optional[CDRRedirectReason] = Field(alias='Redirect reason')
+    redirect_reason: Optional[CDRRedirectReason]
     #: Populated for calls that transfer, hold, wait, and so on. For example:
-    related_reason: Optional[CDRRelatedReason] = Field(alias='Related reason')
+    related_reason: Optional[CDRRelatedReason]
     #: A unique ID for this particular record. This can be used when processing records to aid in deduplication.
-    report_id: Optional[str] = Field(alias='Report ID')
+    report_id: Optional[str]
     #: The time this report was created. Time is in UTC.
-    report_time: Optional[datetime] = Field(alias='Report time')
+    report_time: Optional[datetime]
     #: If present, this field's only reported in Originating records. Route group identifies the route group used for
     #: outbound calls routed via a route group to Premises-based PSTN or an on-prem deployment integrated with Webex
     #: Calling (dial plan or unknown extension).
-    route_group: Optional[str] = Field(alias='Route group')
+    route_group: Optional[str]
     #: The main number for the user's site where the call was made or received.
-    site_main_number: Optional[str] = Field(alias='Site main number')
+    site_main_number: Optional[str]
     #: Site timezone is the offset in minutes from UTC time of the user's timezone.
-    site_timezone: Optional[str] = Field(alias='Site timezone')
+    site_timezone: Optional[str]
     #: If the call is TO or FROM a mobile phone using Webex Go, the Client type will show SIP, and Sub client type
     #: will show MOBILE_NETWORK.
-    sub_client_type: Optional[str] = Field(alias='Sub client type')
+    sub_client_type: Optional[str]
     #: A unique identifier for the user associated with the call. This is a unique identifier across Cisco products.
-    user_uuid: Optional[str] = Field(alias='User UUID')
+    user_uuid: Optional[str]
     #: The user who made or received the call.
-    user: Optional[str] = Field(alias='User')
+    user: Optional[str]
     #: The type of user (user or workspace) that made or received the call. For example:
     #:  * AutomatedAttendantVideo
     #:  * Anchor
@@ -201,15 +221,15 @@ class CDR(ApiModel):
     #:  * CallCenterStandard
     #:  * VoiceXML
     #:  * RoutePoint
-    user_type: Optional[CDRUserType] = Field(alias='User type')
+    user_type: Optional[CDRUserType]
     #: For incoming calls, the telephone number of the user. For outgoing calls, it's the telephone number of the
     #: called party.
-    called_number: Optional[str] = Field(alias='Called number')
+    called_number: Optional[str]
     #: For incoming calls, the telephone number of the calling party. For outgoing calls, it's the telephone number
     #: of the user.
-    calling_number: Optional[str] = Field(alias='Calling number')
+    calling_number: Optional[str]
     #: Location of the report.
-    location: Optional[str] = Field(alias='Location')
+    location: Optional[str]
     #: Dialed digits
     #: The keypad digits as dialed by the user, before pre-translations. This field reports multiple call dial
     #: possibilities:
@@ -222,37 +242,40 @@ class CDR(ApiModel):
     #: also reported, as well as the digits dialed thereafter. Note that when pre-translations have no effect,
     #: the dialed digits field contains the same data as the called number field. This field is only used for
     #: originating (outgoing) Calls and is not available for terminating (incoming) Calls.
-    dialed_digits: Optional[str] = Field(alias='Dialed digits')
+    dialed_digits: Optional[str]
     #: Indicates which party released the call first. The possible values are:
     #:
     #: Local: Used when the local user has released the call first.
     #: Remote: Used when the far end party releases the call first.
     #: Unknown: Used when the call has partial information or is unable to gather enough information about the party
     #: who released the call. It could be because of situations like force lock or because of a session audit failure.
-    releasing_party: Optional[str] = Field(alias='Releasing party')
+    releasing_party: Optional[str]
     #: The Session ID comprises a Universally Unique Identifier (UUID) for each user-agent participating in a call. It
     #: can be used for end-to-end tracking of a SIP session in IP-based multimedia communication. Each call consists of
     #: two UUIDs known as Local Session ID and Remote Session ID.
     #:   * The Remote SessionID is generated from the Terminating user agent.
-    remote_session_id: Optional[str] = Field(alias='Remote SessionID')
+    remote_session_id: Optional[str] = Field(alias='remoteSessionid')
     #: When the call has been redirected one or more times, this field reports the last redirecting number.
     #: Identifies who last redirected the call. Only applies to call scenarios such as transfer, call forwarded calls,
     #: simultaneous rings, etc.
-    redirecting_number: Optional[str] = Field(alias='Redirecting number')
+    redirecting_number: Optional[str]
     #: A unique identifier for the site associated with the call. Unique across Cisco products.
-    site_uuid: Optional[str] = Field(alias='Site UUID')
+    site_uuid: Optional[str]
     #: A unique identifier for the user's department name.
-    department_id: Optional[str] = Field(alias='Department ID')
+    department_id: Optional[str]
     #: Transfer related call ID is used as a call identifier of the other call involved in the transfer. You can share
     #: this ID with Cisco TAC to help them pinpoint parties who are involved during a call transfer.
-    transfer_related_call_id: Optional[str] = Field(alias='Transfer related call ID')
+    transfer_related_call_id: Optional[str]
     #: The user who made or received the call.
     user: Optional[str]
     #: The authorization code admin created for a location or site for users to use. Collected by the
     #: Account/Authorization Codes or Enhanced Outgoing Calling Plan services.
-    authorization_code: Optional[str] = Field(alias='Authorization code')
+    authorization_code: Optional[str]
     #: The device model type the user is using to make or receive the call.
-    model: Optional[str] = Field(alias='Model')
+    model: Optional[str]
+    #: Indicates the time at which the call transfer service was invoked during the call. The invocation time is
+    # shown using the UTC/GMT time zone.
+    call_transfer_time: Optional[datetime]
 
 
 @dataclass(init=False)